fix: 불필요한 훅 제거
